🐛 (services/autoheal): Change the docker socket for the rootless one
In order to reduce potential issues on my docker containers, I will use
a rootless docker to host all services.
In this case, autoheal should use the rootless socket in order to
restart the containers.

NOTE: In order to make things simplier, I use the default user (pi on
Raspbian) to host the docker socket. Of course, this user must be
secured and only myself can connect to this host.