:building_construction: (services/uspd): Move Nut daemon directly on the RPi
Because docker is now rootless, using NUT inside a container is a pain
in ass to do. Instead, I will install directly NUT on the instance using
`PyInfra`.

NOTE: because the RPi uses an overlay FS for the / partition, I need to
handle this case, explaining why I add `with
overlayfs.TemporaryDisableOverlayFS` before installing NUT.

Signed-off-by: Alexandre Nicolaie <xunleii@users.noreply.github.com>