:building_construction: (images): Change how images are built
In order to avoid leaking trascrypt key outside my computer, I don't
want the CI to read any configuration that are encrypted.
However, if I want the CI to test if something is updated, I need to
build the image with some dummy configuration that I know they worked
before.

For this case, I introduced a `BUILD_TYPE` argument on each Dockerfile;
this argument will change which configuration should be used during the
build time.
Currently, there is two types : `live` (for production) and `ci` (for
testing purpose)

Signed-off-by: Alexandre Nicolaie <xunleii@users.noreply.github.com>