:sparkles:(catalog): Migrate AdGuardHome to pulumi
This commit also changes the structure of an application in order to
make it more reusable and easier to maintain.
It also adds the version directly into the package.json file and removes
lincense headers from the spec.ts files (which seems to "broke" vitest).

TODO: add a script to automatically update the version in the
package.json file when Renovate updates the application version.

Signed-off-by: Alexandre Nicolaie <xunleii@users.noreply.github.com>