:technologist:(catalog): Migrate to zot as OCI registry
ZOT registry (https://zotregistry.dev/) seems to be a good
alternative to the current registry we are using. It has an embeded
UI, a garbage collector and retention policies.

However, because is an OCI registry, we need to change the way we
build and push images. This is why this commit changes how we export
images but also, to avlid concurrency issues, why all image have now
an UUID as tag.

Finally, this commit also remove the needs of network host, isolating
a bit more the current DevContainer.

Signed-off-by: Alexandre Nicolaie <xunleii@users.noreply.github.com>