:technologist:(gh): Replace mise by nix/direnv/just
Mise-en-place is a great tool but is a bit more frustrating to use
because it lacks on some feature like templating in all fields.

Also, I trying to learn a bit nix and how it works and I would like to
make this repo more "nix" like. Unfortunately, nix is also a real pain
in ass if some app I would like to use is not available on the Nix Store
(like `czg` or custom built `helm-schema`. Maybe trying DevBox should
help a bit ?

Signed-off-by: Alexandre Nicolaie <xunleii@users.noreply.github.com>